@@ -49,13 +49,13 @@
 		"migrate-etherpads": "node ./migrate-etherpads.js"
 	},
 	"dependencies": {
-		"@feathersjs/authentication": "^4.5.3",
-		"@feathersjs/authentication-local": "^4.5.4",
-		"@feathersjs/configuration": "^4.5.3",
-		"@feathersjs/errors": "^4.5.3",
-		"@feathersjs/express": "^4.5.4",
-		"@feathersjs/feathers": "^4.5.3",
-		"@feathersjs/socketio": "^4.5.4",
+		"@feathersjs/authentication": "^4.5.11",
+		"@feathersjs/authentication-local": "^4.5.11",
+		"@feathersjs/configuration": "^4.5.11",
+		"@feathersjs/errors": "^4.5.11",
+		"@feathersjs/express": "^4.5.11",
+		"@feathersjs/feathers": "^4.5.11",
+		"@feathersjs/socketio": "^4.5.11",
 		"@hpi-schul-cloud/commons": "^1.3.2",
 		"@sentry/node": "^6.2.0",
 		"ajv": "^6.10.2",
@@ -77,9 +77,9 @@
 		"es6-promisify": "^6.0.1",
 		"express": "^4.14.0",
 		"express-openapi-validator": "^4.7.0",
-		"feathers-hooks-common": "^5.0.3",
+		"feathers-hooks-common": "^5.0.5",
 		"feathers-mongoose": "^8.3.1",
-		"feathers-swagger": "^1.1.0",
+		"feathers-swagger": "^1.2.1",
 		"file-type": "^16.2.0",
 		"freeport": "^1.0.5",
 		"html-entities": "^2.1.0",
@@ -91,7 +91,7 @@
 		"mixwith": "^0.1.1",
 		"moment": "^2.19.2",
 		"mongodb-uri": "^0.9.7",
-		"mongoose": "~5.12.0",
+		"mongoose": "~5.12.1",
 		"mongoose-diff-history": "git+https://github.com/schul-cloud/mongoose-diff-history.git",
 		"mongoose-history": "^0.8.0",
 		"mongoose-id-validator": "^0.6.0",
@@ -116,28 +116,37 @@
 		"tiny-async-pool": "^1.2.0",
 		"universal-analytics": "^0.4.17",
 		"urlsafe-base64": "^1.0.0",
-		"uuid": "^8.3.0",
+		"uuid": "^8.3.2",
 		"wait-on": "^5.0.1",
 		"winston": "^3.2.0",
 		"xml2js-es6-promise": "^1.1.1"
 	},
 	"devDependencies": {
 		"@istanbuljs/nyc-config-typescript": "^1.0.1",
-		"@types/express-serve-static-core": "4.17.18",
+		"@types/chai-as-promised": "^7.1.3",
+		"@types/compression": "^1.7.0",
+		"@types/cors": "^2.8.10",
+		"@types/express-serve-static-core": "^4.17.19",
+		"@types/feathersjs__express": "^1.1.8",
 		"@types/mocha": "^8.0.3",
-		"@typescript-eslint/eslint-plugin": "^4.5.0",
-		"@typescript-eslint/parser": "^4.5.0",
+		"@types/mongoose-id-validator": "^0.6.0",
+		"@types/multer": "^1.4.5",
+		"@types/serve-favicon": "^2.5.2",
+		"@types/uuid": "^8.3.0",
+		"@types/xml2js": "^0.4.8",
+		"@typescript-eslint/eslint-plugin": "^4.17.0",
+		"@typescript-eslint/parser": "^4.17.0",
 		"chai": "^4.2.0",
 		"chai-as-promised": "^7.1.1",
 		"chai-http": "^4.2.0",
 		"codecov": "^3.6.5",
 		"copyfiles": "^2.4.0",
-		"eslint": "^7.6.0",
+		"eslint": "^7.22.0",
 		"eslint-config-airbnb-base": "^14.1.0",
-		"eslint-config-prettier": "^8.0.0",
-		"eslint-plugin-import": "^2.18.2",
-		"eslint-plugin-prettier": "^3.1.4",
-		"eslint-plugin-promise": "^4.2.1",
+		"eslint-config-prettier": "^8.1.0",
+		"eslint-plugin-import": "^2.22.1",
+		"eslint-plugin-prettier": "^3.3.1",
+		"eslint-plugin-promise": "^4.3.1",
 		"jwt-decode": "^3.1.2",
 		"mocha": "^8.0.1",
 		"mockery": "^2.0.0",
