@@ -54,6 +54,13 @@ module.exports = {
 		mocha: true,
 		browser: true,
 	},
+	settings: {
+		'import/resolver': {
+			node: {
+				extensions: ['.js', '.ts'],
+			},
+		},
+	},
 	overrides: [
 		{
 			files: ['*.test.js', '*.test.ts'],
@@ -65,8 +72,20 @@ module.exports = {
 		{
 			files: ['*.ts'],
 			parser: '@typescript-eslint/parser',
-			plugins: ['@typescript-eslint'],
-			extends: ['plugin:@typescript-eslint/recommended', 'prettier/@typescript-eslint', 'plugin:prettier/recommended'],
+			extends: [
+				'plugin:@typescript-eslint/recommended', // Uses the recommended rules from the @typescript-eslint/eslint-plugin
+				'plugin:import/typescript',
+				'plugin:prettier/recommended', // Enables eslint-plugin-prettier and eslint-config-prettier. This will display prettier errors as ESLint errors. Make sure this is always the last configuration in the extends array.
+			],
+			parserOptions: {
+				ecmaVersion: 2020, // Allows for the parsing of modern ECMAScript features
+				sourceType: 'module', // Allows for the use of imports
+			},
+			rules: {
+				'import/extensions': ['error', 'never'], // We not want extensions required within of import statements
+				'import/prefer-default-export': 'off', // JS can not import default from TS
+				'import/no-extraneous-dependencies': ['error', { devDependencies: ['**/*.test.{js,ts}'] }], // We allow dev dependencies to be used in test files only
+			},
 		},
 	],
 };
